+++ src/Tuple.java
@@ -1,15 +1,15 @@
 
 
 public class Tuple {
-    int x1;
-    int x2;
+    long x1;
+    long x2;
 
-    public Tuple(int x1, int x2)
+    public Tuple(long x1, long x2)
     {
         this.x1 = x1;
         this.x2 = x2;
     }
 
-    public int getX(){ return x1;}
-    public int getY(){ return x2;}
+    public long getX(){ return x1;}
+    public long getY(){ return x2;}
 }
